+++ CHANGELOG.md
@@ -16,7 +16,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - [CSE] Added support for TimeSeries/TimeSeriesInstance resource types.
 - [CSE] Added support for *ctm* (currentTime) attribute for &lt;CSEBase> resource type.
 - [CSE] Added wildcard (\*) support for &lt;ACP>'s *acr/acor* originators.
-- [WEB] Added oauth2 authorization support for the proxied CSE (for the stand-alone web UI).
+- [WEB] Added OAuth2 authorization support for the proxied CSE (for the stand-alone web UI).
+- [WEB] Added opening the web UI in a browser on startup (for the stand-alone web UI).
 
 ### Changed
 - [CSE] Relaxed validation for float in attributes and arguments. Integer are now accepted as well.